fun: 补充注释、添加自定义validator校验时间

+++ src/main/java/com/example/smarter/entity/User.java
@@ -13,11 +13,35 @@ import java.time.LocalDateTime;
 @Data
 @AllArgsConstructor
 public class User {
-    private Long id;
+
+    /**
+     * 用户 id
+     */
+    private Integer id;
+
+    /**
+     * 用户名
+     */
     private String username;
+
+    /**
+     * 用户密码哈希值
+     */
     private String password;
+
+    /**
+     * 用户名联系电话
+     */
     private String phone;
+
+    /**
+     * 用户创建时间
+     */
     private LocalDateTime createTime;
+
+    /**
+     * 用户修改时间
+     */
     private LocalDateTime updateTime;
 
     public User() {}
